team comp logic is done!
v1 of the logic to find team comp is done and prints to cli currently. there is also now a page to save champ roles from the csv rather than manually nuking and re-creating it after changes. we will also clean up how picky it is about the team comp score and how many champs we give it to work over